Clean up. Roundtrip Log tests
