+++ src/HdrHistogram.UnitTests/HistogramEncodingTest.cs
@@ -8,7 +8,6 @@
  * https://github.com/HdrHistogram/HdrHistogram
  */
 
-using System.Linq;
 using HdrHistogram.Encoding;
 using HdrHistogram.Utilities;
 using NUnit.Framework;
@@ -28,7 +27,7 @@ namespace HdrHistogram.UnitTests
             var source = Create(HighestTrackableValue, 3);
             Load(source);
             var result = EncodeDecode(source);
-            AssertAreEqual(source, result);
+            HistogramAssert.AreEqual(source, result);
         }
 
         [Test]
@@ -37,7 +36,7 @@ namespace HdrHistogram.UnitTests
             var source = Create(HighestTrackableValue, 3);
             Load(source);
             var result = CompressedEncodeDecode(source);
-            AssertAreEqual(source, result);
+            HistogramAssert.AreEqual(source, result);
         }
 
         [Test]
@@ -46,7 +45,7 @@ namespace HdrHistogram.UnitTests
             var source = Create(HighestTrackableValue, 3);
             LoadFullRange(source);
             var result = EncodeDecode(source);
-            AssertAreEqual(source, result);
+            HistogramAssert.AreEqual(source, result);
         }
 
         [Test]
@@ -55,7 +54,7 @@ namespace HdrHistogram.UnitTests
             var source = Create(HighestTrackableValue, 3);
             LoadFullRange(source);
             var result = CompressedEncodeDecode(source);
-            AssertAreEqual(source, result);
+            HistogramAssert.AreEqual(source, result);
         }
 
         private static T EncodeDecode(T source)
@@ -91,21 +90,6 @@ namespace HdrHistogram.UnitTests
             source.RecordValue(HighestTrackableValue);
         }
 
-        private static void AssertAreEqual(HistogramBase expected, HistogramBase actual)
-        {
-            Assert.AreEqual(expected.GetType(), actual.GetType());
-            Assert.AreEqual(expected.TotalCount, actual.TotalCount);
-            Assert.AreEqual(expected.StartTimeStamp, actual.StartTimeStamp);
-            Assert.AreEqual(expected.EndTimeStamp, actual.EndTimeStamp);
-            Assert.AreEqual(expected.LowestTrackableValue, actual.LowestTrackableValue);
-            Assert.AreEqual(expected.HighestTrackableValue, actual.HighestTrackableValue);
-            Assert.AreEqual(expected.NumberOfSignificantValueDigits, actual.NumberOfSignificantValueDigits);
-            var expectedValues = expected.AllValues().ToArray();
-            var actualValues = actual.AllValues().ToArray();
-
-            //TODO: We are 
-            CollectionAssert.AreEqual(expectedValues, actualValues, HistogramIterationValueComparer.Instance);
-        }
     }
 
     [TestFixture]
