+++ src/HdrHistogram.UnitTests/HistogramLogReaderWriterTests.cs
@@ -35,7 +35,56 @@ namespace HdrHistogram.UnitTests
             }
         }
 
-        [TestCase("Resources\\jHiccup-2.0.7S.logV2.hlog")]
+        [Test]
+        public void CanRoundTripSingleHsitogram()
+        {
+            var histogram = CreatePopulatedHistogram(1000);
+            var startTimeWritten = DateTime.Now;
+            var endTimeWritten = startTimeWritten.AddMinutes(30);
+            
+            histogram.StartTimeStamp = (long)(startTimeWritten.SecondsSinceUnixEpoch()*1000L);
+            histogram.EndTimeStamp = (long)(endTimeWritten.SecondsSinceUnixEpoch()*1000L);
+
+            var data = WriteLog(startTimeWritten, histogram);
+            var actualHistograms = ReadHistograms(data);
+
+            Assert.AreEqual(1, actualHistograms.Length);
+            HistogramAssert.AreEqual(histogram, actualHistograms.Single());
+        }
+
+        private static HistogramBase[] ReadHistograms(byte[] data)
+        {
+            HistogramBase[] actualHistograms;
+            using (var readerStream = new MemoryStream(data))
+            {
+                var reader = new HistogramLogReader(readerStream);
+                actualHistograms = reader.ReadHistograms().ToArray();
+            }
+            return actualHistograms;
+        }
+
+        private static byte[] WriteLog(DateTime startTimeWritten, LongHistogram histogram)
+        {
+            byte[] data;
+            using (var writerStream = new MemoryStream())
+            {
+                var writer = new HistogramLogWriter(writerStream);
+                writer.Write(startTimeWritten, histogram);
+                data = writerStream.ToArray();
+            }
+            return data;
+        }
+
+        private static LongHistogram CreatePopulatedHistogram(long multiplier)
+        {
+            var histogram = new LongHistogram(3600L * 1000 * 1000, 3);
+            for (int i = 0; i < 10000; i++)
+            {
+                histogram.RecordValue(i * multiplier);
+            }
+            return histogram;
+        }
+
         [TestCase("Resources\\jHiccup-2.0.7S.logV2.hlog")]
         public void CanReadv2Logs(string logPath)
         {
