+++ src/HdrHistogram/Encoding/IEncoder.cs
@@ -1,0 +1,9 @@
+using HdrHistogram.Utilities;
+
+namespace HdrHistogram.Encoding
+{
+    public interface IEncoder
+    {
+        int Encode(IRecordedData data, ByteBuffer buffer);
+    }
+}
