+++ src/HdrHistogram/Histogram.cs
@@ -250,58 +250,4 @@ namespace HdrHistogram
             }
         }
     }
-
-    public interface IHeader
-    {
-        int Cookie { get; }
-        int PayloadLengthInBytes { get; }
-        int NormalizingIndexOffset { get; }
-        int NumberOfSignificantValueDigits { get; }
-        long LowestTrackableUnitValue { get; }
-        long HighestTrackableValue { get; }
-        double IntegerToDoubleValueConversionRatio { get; }
-    }
-
-    sealed class V1Header : IHeader
-    {
-        public V1Header(int cookie, ByteBuffer buffer)
-        {
-            Cookie = cookie;
-            PayloadLengthInBytes = buffer.GetInt();
-            NormalizingIndexOffset = buffer.GetInt();
-            NumberOfSignificantValueDigits = buffer.GetInt();
-            LowestTrackableUnitValue = buffer.GetLong();
-            HighestTrackableValue = buffer.GetLong();
-            IntegerToDoubleValueConversionRatio = buffer.GetDouble();
-        }
-
-        public int Cookie { get; }
-        public int PayloadLengthInBytes { get; }
-        public int NormalizingIndexOffset { get; }
-        public int NumberOfSignificantValueDigits { get; }
-        public long LowestTrackableUnitValue { get; }
-        public long HighestTrackableValue { get; }
-        public double IntegerToDoubleValueConversionRatio { get; }
-    }
-    sealed class V0Header : IHeader
-    {
-        public V0Header(int cookie, ByteBuffer buffer)
-        {
-            Cookie = cookie;
-            NumberOfSignificantValueDigits = buffer.GetInt();
-            LowestTrackableUnitValue = buffer.GetLong();
-            HighestTrackableValue = buffer.GetLong();
-            buffer.GetLong(); // Discard totalCount field in V0 header.
-            PayloadLengthInBytes = int.MaxValue;
-            IntegerToDoubleValueConversionRatio = 1.0;
-            NormalizingIndexOffset = 0;
-        }
-        public int Cookie { get; }
-        public int PayloadLengthInBytes { get; }
-        public int NormalizingIndexOffset { get; }
-        public int NumberOfSignificantValueDigits { get; }
-        public long LowestTrackableUnitValue { get; }
-        public long HighestTrackableValue { get; }
-        public double IntegerToDoubleValueConversionRatio { get; }
-    }
 }
